requirements.txt
# ===============================================
# DJANGO FACE RECOGNITION ATTENDANCE SYSTEM
# Requirements for Production & Development
# ===============================================

# Core Django Framework
Django==5.2.8
django-widget-tweaks==1.5.0

# Database Drivers
mysqlclient==2.2.7              # MySQL driver for production
mysql-connector-python==9.5.0   # Alternative MySQL connector

# Computer Vision & Face Recognition
opencv-python==4.12.0.88         # OpenCV for image processing
opencv-contrib-python==4.11.0.86 # OpenCV additional modules
dlib==20.0.0                    # Facial landmarks (versión que funciona en este proyecto)
face-recognition==1.3.0          # Face recognition library
face_recognition_models==0.3.0   # Pre-trained face models

# Machine Learning & AI
mediapipe==0.10.21               # Hand gesture detection
numpy==1.26.4                    # Numerical computing
scipy==1.16.3                    # Scientific computing
pandas==2.3.3                    # Data manipulation

# Image Processing & Graphics
Pillow==12.0.0                   # Image processing library
matplotlib==3.10.7               # Plotting for debugging

# Core Python Dependencies
cffi==2.23                       # C Foreign Function Interface
pycparser==2.23                  # C parser for cffi
packaging==25.0                  # Package utilities
six==1.17.0                      # Python 2/3 compatibility

# Date & Time
python-dateutil==2.9.0.post0    # Date utilities
pytz==2025.2                     # Timezone handling

# Utilities
colorama==0.4.6                  # Colored terminal output
click==8.3.0                     # CLI utilities

# ===============================================
# INSTALLATION INSTRUCTIONS:
# ===============================================
# 1. Install Python 3.8-3.11 (recommended: 3.11)
# 2. Install Visual Studio Build Tools
# 3. Install CMake: pip install cmake
# 4. Create virtual environment: python -m venv .venv
# 5. Activate: .venv\Scripts\activate (Windows)
# 6. Install: pip install -r requirements.txt
# 
# TROUBLESHOOTING:
# - For dlib issues: Use dlib==20.0.0 (versión probada y funcional)
# - If dlib fails: pip install cmake, then retry
# - For MySQL: Install MySQL Client libraries
# ===============================================